Conflicts:
	SP2/Debug/LiZhaoyuan1.exe
	SP2/Debug/LiZhaoyuan1.ilk
	SP2/Debug/LiZhaoyuan1.pdb
	SP2/LiZhaoyuan1/Debug/Application.obj
	SP2/LiZhaoyuan1/Debug/Camera3.obj
	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.log
	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.tlog/CL.read.1.tlog
	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.tlog/CL.write.1.tlog
	SP2/LiZhaoyuan1/Debug/LiZhaoyuan1.tlog/LiZhaoyuan1.lastbuildstate
	SP2/LiZhaoyuan1/Debug/SceneGalaxy.obj
	SP2/LiZhaoyuan1/Debug/SceneSoraJewel.obj
	SP2/LiZhaoyuan1/Debug/vc120.idb
	SP2/LiZhaoyuan1/Debug/vc120.pdb
	SP2/LiZhaoyuan1/Source/Camera3.cpp

+++ SP2/LiZhaoyuan1/Source/SceneSoraJewel.h
@@ -137,7 +137,7 @@ private:
 	Light light[2];
 	bool enableLight = false;
 	float LSPEED = 10.f;
-	double fps = 0;
+	int fps = 0;
 
 	void renderMesh(Mesh *mesh, bool enableLight);
 	void lighting();
